feat: convert decompiling to gradle worker
