+++ src/main/kotlin/lol/koblizek/composer/actions/DecompileAction.kt
@@ -1,73 +1,22 @@
 package lol.koblizek.composer.actions
 
-import lol.koblizek.composer.ComposerPlugin
-import org.apache.commons.io.FileUtils
+import lol.koblizek.composer.util.VineflowerWorker
 import org.gradle.api.Project
-import org.jetbrains.java.decompiler.main.Fernflower
-import org.jetbrains.java.decompiler.main.decompiler.DirectoryResultSaver
-import org.jetbrains.java.decompiler.main.extern.IFernflowerLogger
-import org.jetbrains.java.decompiler.main.extern.IFernflowerPreferences
-import java.io.File
+import org.gradle.workers.WorkerExecutor
+import javax.inject.Inject
 
-class DecompileAction : Action() {
+abstract class DecompileAction : Action() {
+    @Inject
+    public abstract fun getWorkerExecutor(): WorkerExecutor;
+
     override fun run(project: Project) {
-        if (!ComposerPlugin.isConfigInitialized() || ComposerPlugin.config.decompilationSource == null) {
-            println("Cannot decompile: Missing runtimeConfig block or decompilationSource parameter")
-            println("Run cleanUp task and rebuild the project once you fix the problem")
-            return
-        }
-
-        val props: MutableMap<String, Any> = HashMap(IFernflowerPreferences.DEFAULTS)
-        props[IFernflowerPreferences.DECOMPILE_GENERIC_SIGNATURES] = "1"
-        props[IFernflowerPreferences.REMOVE_SYNTHETIC] = "1"
-        props[IFernflowerPreferences.INCLUDE_ENTIRE_CLASSPATH] = "1"
-        props[IFernflowerPreferences.PATTERN_MATCHING] = "1"
-        props[IFernflowerPreferences.TERNARY_CONDITIONS] = "1"
-        props[IFernflowerPreferences.THREADS] = (Runtime.getRuntime().availableProcessors() / 2).toString()
-
-        val dir = File(ComposerPlugin.config.decompilationSource!!)
-
-        val fernFlower = Fernflower(
-            DirectoryResultSaver(dir),
-            props,
-            Logger()
-        )
-        fernFlower.addSource(DeobfuscateAction().temporaryDir.resolve("server-deobf.jar"))
-        project.configurations.getByName("compileClasspath").files.forEach {
-            fernFlower.addLibrary(it)
-        }
-        ComposerPlugin.config.toRemove.forEach {
-            val file = dir.toPath().resolve(it).toFile()
-            if (file.exists()) {
-                if (file.isDirectory) file.deleteRecursively()
-                else file.delete()
+        val queue = getWorkerExecutor().processIsolation {
+            it.forkOptions { java ->
+                java.maxHeapSize = "2G"
             }
         }
-        if (ComposerPlugin.config.resourcesSource != null) {
-            ComposerPlugin.config.resources.forEach {
-                FileUtils.moveToDirectory(File(it), File(ComposerPlugin.config.resourcesSource!!), true)
-            }
+        queue.submit(VineflowerWorker::class.java) {
+            it.project = project
         }
-        ComposerPlugin.config.moveToRoot.forEach {
-            val root = ComposerPlugin.config.decompilationSource
-            val file = File(root!!).resolve(it)
-            FileUtils.moveToDirectory(file, File(root), true)
-        }
-        fernFlower.decompileContext()
-    }
-
-    class Logger : IFernflowerLogger() {
-        override fun writeMessage(message: String, severity: Severity) {
-            if (severity.ordinal >= Severity.WARN.ordinal) {
-                println(message)
-            }
-        }
-
-        override fun writeMessage(message: String, severity: Severity, t: Throwable) {
-            if (severity.ordinal >= Severity.WARN.ordinal) {
-                println(message)
-            }
-        }
-
     }
 }
